[KARAF-2113] In the RepositoryEventHandler, check the local status of a features repository before modifying the features service
git-svn-id: https://svn.apache.org/repos/asf/karaf/cellar/trunk@1448693 13f79535-47bb-0310-9956-ffa450edef68

+++ features/src/main/java/org/apache/karaf/cellar/features/RepositoryEventHandler.java
@@ -63,11 +63,19 @@ public class RepositoryEventHandler extends FeaturesSupport implements EventHand
         try {
             // TODO check if isAllowed
             if (RepositoryEvent.EventType.RepositoryAdded.equals(type)) {
-                LOGGER.debug("CELLAR FEATURES: adding repository URI {}", uri);
-                featuresService.addRepository(new URI(uri));
+                if (!isRepositoryRegisteredLocally(uri)) {
+                    LOGGER.debug("CELLAR FEATURES: adding repository URI {}", uri);
+                    featuresService.addRepository(new URI(uri));
+                } else {
+                    LOGGER.debug("CELLAR FEATURES: repository URI {} is already registered locally");
+                }
             } else {
-                LOGGER.debug("CELLAR FEATURES: removing repository URI {}", uri);
-                featuresService.removeRepository(new URI(uri));
+                if (isRepositoryRegisteredLocally(uri)) {
+                    LOGGER.debug("CELLAR FEATURES: removing repository URI {}", uri);
+                    featuresService.removeRepository(new URI(uri));
+                } else {
+                    LOGGER.debug("CELLAR FEATURES: repository URI {} is not registered locally");
+                }
             }
         } catch (Exception e) {
             LOGGER.error("CELLAR FEATURES: failed to add/remove repository URL {}", uri, e);
